Refactored Player Page to Reflect Figma
