+++ src/apps/admin/pages/MusicQueue.tsx
@@ -1,28 +1,23 @@
 import { useSelector } from 'react-redux'
 import { TrackList } from 'src/components'
-import {
-  selectCurrentJukebox,
-  selectCurrentMembership,
-  selectNextTracks,
-  selectUser
-} from 'src/store'
+import { selectNextTracks } from 'src/store'
 import './MusicQueue.scss'
+import { TrackModifyContext } from './trackContext'
 
 export const MusicQueue = () => {
   const nextTracks = useSelector(selectNextTracks)
-  const currentJukebox = useSelector(selectCurrentJukebox)
-  const currentUser = useSelector(selectUser)
-  const currentMembership = useSelector(selectCurrentMembership)
 
   return (
     <>
-      <div className="music-queue_container">
-        <div className="next_up_container">
-          <h1>Next Up</h1>
+      <TrackModifyContext.Provider value={true}>
+        <div className="music-queue_container">
+          <div className="next_up_container">
+            <h1>Next Up</h1>
 
-          <TrackList tracks={nextTracks} />
+            <TrackList tracks={nextTracks} />
+          </div>
         </div>
-      </div>
+      </TrackModifyContext.Provider>
     </>
   )
 }
