+++ src/apps/admin/pages/Player.scss
@@ -1,0 +1,78 @@
+.spotify-player-container {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  margin-left: 1vh;
+}
+
+.spotify-player-title {
+  @include font-display('xs');
+  color: var(--color-primary-contrast);
+  margin-bottom: 2%;
+  margin-top: 2%;
+  justify-items: center;
+}
+
+.left-container {
+  margin-right: 18%;
+}
+
+.right-container {
+  padding-top: 1%;
+}
+
+.spotify-play-desc {
+  margin-bottom: 8vh;
+}
+
+.spotify-song-title {
+  @include font-title('lg');
+  margin-bottom: 1vh;
+}
+
+.spotify-song-author {
+  @include font-label('md');
+  margin-bottom: 2vh;
+}
+
+.audio-container {
+  margin-bottom: 6vh;
+}
+
+.detail-container {
+  margin-bottom: 4%;
+}
+
+.next-track-container {
+  margin-top: 8%;
+}
+
+.session-container {
+  margin-bottom: 10%;
+}
+
+.connect-button-container {
+  margin-top: 5%;
+}
+
+.connect-button {
+  border-radius: 15px !important;
+}
+
+.debugInfo {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  min-width: 30vw;
+  margin-bottom: 2rem;
+}
+
+.switchAccounts {
+  margin-top: 1rem;
+  margin-bottom: 3rem;
+}
+
+.spotify-accounts{
+  display: flex;
+  flex-direction: column;
+  margin-top: 2rem;
+}
