+++ src/apps/admin/pages/trackContext.tsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react'
+
+export const TrackModifyContext = createContext<boolean>(false)
