+++ src/styles/components/_components.scss
@@ -5,5 +5,7 @@ $component-colors: (
   'warning',
   'success',
   'disabled',
-  'neutral'
+  'neutral',
+  'success-contrast-on',
+  'success-contrast',
 );
